refactor(GN-150): 유틸 함수 분리
